Make application of W3C CLA to proposals clear
In order to handle proposals that the group wants to consider and work with, but has not formally adopted, we are adding specific text to make it clear that any work done in this repository is specifically for the purposes of generating specifications or documents for PATCG and as such is covered by the W3C CLA. 

Added: 

This is the "proposals" repository of the PATCG. As a result, any documents that enter this repository should be considered as under consideration for specification work or supporting that work. Work done within this repository is therefore fully covered under the [W3C CLA](https://www.w3.org/community/about/agreements/cla/) for the purpose of generating documents and specifications as the standard output of PATCG. 

Documents in the repository are in one of two stages, draft or final. The presence of a proposal in this repository implies that it is at least in a draft state and under consideration by this Community Group but does not mean it has been officially set on the specification track for a final specification output. Some documents may never leave the draft state and any document will be specially marked as in its final state when it achieves that state.