Adding Work Process section
Added: 

With the approval of the chairs and group, documents may enter this repository for evaluation and refinement as part of the process of working on proposals and getting them adopted. In order to support this process, please create the applicable new documents in their own folder so that each document in progress has a folder in this repository. In doing so these documents become governed by the [W3C Community License Agreement (CLA)](https://www.w3.org/community/about/agreements/cla/).